--- a/parse_json_amd64.go
+++ b/parse_json_amd64.go
@@ -46,6 +46,7 @@ func (pj *internalParsedJson) initialize(size int) {
 		pj.containingScopeOffset = make([]uint64, 0, maxdepth)
 	}
 	pj.containingScopeOffset = pj.containingScopeOffset[:0]
+	pj.indexesChan = indexChan{}
 }
 
 func (pj *internalParsedJson) parseMessage(msg []byte, ndjson bool) (err error) {
@@ -63,7 +64,9 @@ func (pj *internalParsedJson) parseMessage(msg []byte, ndjson bool) (err error) 
 	// Make the capacity of the channel smaller than the number of slots.
 	// This way the sender will automatically block until the consumer
 	// has finished the slot it is working on.
-	pj.indexChans = make(chan indexChan, indexSlots-2)
+	if pj.indexChans == nil {
+		pj.indexChans = make(chan indexChan, indexSlots-2)
+	}
 	pj.buffersOffset = ^uint64(0)
 
 	var errStage1 error
@@ -74,29 +77,43 @@ func (pj *internalParsedJson) parseMessage(msg []byte, ndjson bool) (err error) 
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
-			if !pj.unifiedMachine(pj.Message) {
+			if !pj.unifiedMachine() {
 				err = errors.New("Bad parsing while executing stage 2")
-				// drain the channel until empty
-				for range pj.indexChans {
+				// Keep consuming...
+				for idx := range pj.indexChans {
+					if idx.index == -1 {
+						break
+					}
 				}
 			}
 		}()
-		if !pj.findStructuralIndices(pj.Message) {
+		if !pj.findStructuralIndices() {
 			errStage1 = errors.New("Failed to find all structural indices for stage 1")
 		}
 		wg.Wait()
 	} else {
-		if !pj.findStructuralIndices(pj.Message) {
+		if !pj.findStructuralIndices() {
 			// drain the channel until empty
-			for range pj.indexChans {
+			for idx := range pj.indexChans {
+				if idx.index == -1 {
+					break
+				}
 			}
 			return errors.New("Failed to find all structural indices for stage 1")
 		}
-		if !pj.unifiedMachine(pj.Message) {
+		if !pj.unifiedMachine() {
 			// drain the channel until empty
-			for range pj.indexChans {
+			for {
+				select {
+				case idx := <-pj.indexChans:
+					if idx.index == -1 {
+						return errors.New("Bad parsing while executing stage 2")
+					}
+					// Already drained.
+				default:
+					return errors.New("Bad parsing while executing stage 2")
+				}
 			}
-			return errors.New("Bad parsing while executing stage 2")
 		}
 		return nil
 	}